karma: diverge old and new
karma     = Original Karma plugin
karma_adv = Karma with Stats
