+++ hamper/plugins/karma.py
@@ -1,15 +1,10 @@
 import re
-from collections import Counter, defaultdict
-from datetime import datetime
+from collections import defaultdict
 
 from hamper.interfaces import ChatCommandPlugin, Command
 from hamper.utils import ude, uen
 
-import pytz
-from pytz import timezone
-from pytz.exceptions import UnknownTimeZoneError
-
-from sqlalchemy import Column, DateTime, Integer, String, func
+from sqlalchemy import Column, Integer, String
 from sqlalchemy.ext.declarative import declarative_base
 
 SQLAlchemyBase = declarative_base()
@@ -22,6 +17,10 @@ class Karma(ChatCommandPlugin):
     Hamper will look for lines that end in ++ or -- and modify that user's
     karma value accordingly
 
+    !karma --top: shows (at most) the top 5
+    !karma --bottom: shows (at most) the bottom 5
+    !karma <username>: displays the karma for a given user
+
     NOTE: The user is just a string, this really could be anything...like
     potatoes or the infamous cookie clicker....
     """
@@ -30,17 +29,12 @@ class Karma(ChatCommandPlugin):
 
     priority = -2
 
-    short_desc = ("karma - Give positive or negative karma. Where you see"
-                  " !karma, !score will work as well")
-    long_desc = ("username++ - Give karma\n"
-                 "username-- - Take karma\n"
-                 "!karma --top - Show the top 5 karma earners\n"
-                 "!karma --bottom - Show the bottom 5 karma earners\n"
-                 "!karma --giver or --taker - Show who's given the most"
-                 " positive or negative karma\n"
-                 "!karma --when-positive or --when-negative "
-                 " - Show when people are the most positive or negative\n"
-                 "!karma username - Show the user's karma count\n")
+    short_desc = 'karma/score - Give or take karma from someone'
+    long_desc = ('username++ - Give karma\n'
+                 'username-- - Take karma\n'
+                 '!karma --top - Show the top 5 karma earners\n'
+                 '!karma --bottom - Show the bottom 5 karma earners\n'
+                 '!karma username - Show the user\'s karma count\n')
 
     gotta_catch_em_all = r"""# 3 or statement
                              (
@@ -67,15 +61,6 @@ class Karma(ChatCommandPlugin):
         self.db = loader.db
         SQLAlchemyBase.metadata.create_all(self.db.engine)
 
-        # Config
-        config = loader.config.get("karma", {})
-        self.timezone = config.get('timezone', 'UTC')
-        try:
-            self.tzinfo = timezone(self.timezone)
-        except UnknownTimeZoneError:
-            self.tzinfo = timezone('UTC')
-            self.timezone = 'UTC'
-
     def message(self, bot, comm):
         """
         Check for strings ending with 2 or more '-' or '+'
@@ -91,12 +76,9 @@ class Karma(ChatCommandPlugin):
             karmas = self.modify_karma(words)
             # Notify the users they can't modify their own karma
             if comm['user'] in karmas.keys():
-                if karmas[comm['user']] <= 0:
-                    bot.reply(comm, "Don't be so hard on yourself.")
-                else:
-                    bot.reply(comm, "Tisk, tisk, no up'ing your own karma.")
+                bot.reply(comm, "Nice try, no modifying your own karma")
             # Commit karma changes to the db
-            self.update_db(comm["user"], karmas)
+            self.update_db(karmas, comm['user'])
 
     def modify_karma(self, words):
         """
@@ -130,23 +112,26 @@ class Karma(ChatCommandPlugin):
                     k[word] += change
         return k
 
-    def update_db(self, giver, receiverkarma):
+    def update_db(self, userkarma, username):
         """
-        Record a the giver of karma, the receiver of karma, and the karma
-        amount. Typically the count will be 1, but it can be any positive or
-        negative integer.
+        Change the users karma by the karma amount (either 1 or -1)
         """
 
-        for receiver in receiverkarma:
-            if receiver != giver:
-                urow = KarmaStatsTable(ude(giver), ude(receiver),
-                                       receiverkarma[receiver])
+        kt = self.db.session.query(KarmaTable)
+        for user in userkarma:
+            if user != username:
+                # Modify the db accourdingly
+                urow = kt.filter(KarmaTable.user == ude(user)).first()
+                # If the user doesn't exist, create it
+                if not urow:
+                    urow = KarmaTable(ude(user))
+                urow.kcount += userkarma[user]
                 self.db.session.add(urow)
         self.db.session.commit()
 
     class KarmaList(Command):
         """
-        Return the highest or lowest 5 receivers of karma
+        Return the top or bottom 5
         """
 
         regex = r'^(?:score|karma) --(top|bottom)$'
@@ -154,51 +139,19 @@ class Karma(ChatCommandPlugin):
         LIST_MAX = 5
 
         def command(self, bot, comm, groups):
-            # Let the database restrict the amount of rows we get back.
-            # We can then just deal with a few rows later on
-            session = bot.factory.loader.db.session
-            kcount = func.sum(KarmaStatsTable.kcount).label('kcount')
-            kts = session.query(KarmaStatsTable.receiver, kcount) \
-                         .group_by(KarmaStatsTable.receiver)
+            users = bot.factory.loader.db.session.query(KarmaTable)
+            user_count = users.count()
+            top = self.LIST_MAX if user_count >= self.LIST_MAX else user_count
 
-            # For legacy support
-            classic = session.query(KarmaTable)
-
-            # Counter for sorting and updating data
-            counter = Counter()
-
-            if kts.count() or classic.count():
-                # We should limit the list of users to at most self.LIST_MAX
-                if groups[0] == 'top':
-                    classic_q = classic.order_by(KarmaTable.kcount.desc())\
-                                       .limit(self.LIST_MAX).all()
-                    query = kts.order_by(kcount.desc())\
-                               .limit(self.LIST_MAX).all()
-
-                    counter.update(dict(classic_q))
-                    counter.update(dict(query))
-                    snippet = counter.most_common(self.LIST_MAX)
-                elif groups[0] == 'bottom':
-                    classic_q = classic.order_by(KarmaTable.kcount)\
-                                       .limit(self.LIST_MAX).all()
-                    query = kts.order_by(kcount)\
-                               .limit(self.LIST_MAX).all()
-                    counter.update(dict(classic_q))
-                    counter.update(dict(query))
-                    snippet = reversed(counter.most_common(self.LIST_MAX))
-                else:
+            if top:
+                show = (KarmaTable.kcount.desc() if groups[0] == 'top'
+                        else KarmaTable.kcount)
+                for user in users.order_by(show)[0:top]:
                     bot.reply(
-                        comm, r'Something went wrong with karma\'s regex'
-                    )
-                    return
-
-                for rec in snippet:
-                    bot.reply(
-                        comm, '%s\x0f: %d' % (uen(rec[0]), rec[1]),
-                        encode=False
+                        comm, str('%s\x0f: %d' % (user.user, user.kcount))
                     )
             else:
-                bot.reply(comm, 'No one has any karma yet :-(')
+                bot.reply(comm, r'No one has any karma yet :-(')
 
     class UserKarma(Command):
         """
@@ -206,185 +159,39 @@ class Karma(ChatCommandPlugin):
         """
 
         # !karma <username>
-        regex = r'^(?:score|karma)(?:\s+([^-].*))?$'
+        regex = r'^(?:score|karma)\s+([^-].*)$'
 
         def command(self, bot, comm, groups):
-            # The receiver (or in old terms, user) of the karma being tallied
-            receiver = groups[0]
-            if receiver is None:
-                reciever = comm['user']
-            receiver = ude(reciever.strip().lower())
+            # Play nice when the user isn't in the db
+            kt = bot.factory.loader.db.session.query(KarmaTable)
+            thing = ude(groups[0].strip().lower())
+            user = kt.filter(KarmaTable.user == thing).first()
 
-            # Manage both tables
-            sesh = bot.factory.loader.db.session
-
-            # Old Table
-            kt = sesh.query(KarmaTable)
-            user = kt.filter(KarmaTable.user == receiver).first()
-
-            # New Table
-            kst = sesh.query(KarmaStatsTable)
-            kst_list = kst.filter(KarmaStatsTable.receiver == receiver).all()
-
-            # The total amount of karma from both tables
-            total = 0
-
-            # Add karma from the old table
             if user:
-                total += user.kcount
-
-            # Add karma from the new table
-            if kst_list:
-                for row in kst_list:
-                    total += row.kcount
-
-            # Pluralization
-            points = "points"
-            if total == 1 or total == -1:
-                points = "point"
-
-            # Send the message
-            bot.reply(
-                comm, '%s has %d %s' % (uen(receiver), total, points),
-                encode=False
-            )
-
-    class KarmaGiver(Command):
-        """
-        Identifies the person who gives the most karma
-        """
-
-        regex = r'^(?:score|karma) --(giver|taker)$'
-
-        def command(self, bot, comm, groups):
-            kt = bot.factory.loader.db.session.query(KarmaStatsTable)
-            counter = Counter()
-
-            if groups[0] == 'giver':
-                positive_karma = kt.filter(KarmaStatsTable.kcount > 0)
-                for row in positive_karma:
-                    counter[row.giver] += row.kcount
-
-                m = counter.most_common(1)
-                most = m[0] if m else None
-                if most:
-                    bot.reply(
-                        comm,
-                        '%s has given the most karma (%d)' %
-                        (uen(most[0]), most[1])
-                    )
-                else:
-                    bot.reply(
-                        comm,
-                        'No positive karma has been given yet :-('
-                    )
-            elif groups[0] == 'taker':
-                negative_karma = kt.filter(KarmaStatsTable.kcount < 0)
-                for row in negative_karma:
-                    counter[row.giver] += row.kcount
-
-                m = counter.most_common()
-                most = m[-1] if m else None
-                if most:
-                    bot.reply(
-                        comm,
-                        '%s has given the most negative karma (%d)' %
-                        (uen(most[0]), most[1])
-                    )
-                else:
-                    bot.reply(
-                        comm,
-                        'No negative karma has been given yet'
-                    )
-
-    class MostActive(Command):
-        """
-        Least/Most active hours of karma giving/taking
-
-        This will now look in the config for a timezone to use when displaying
-        the hour.
-
-        Example
-        Karma:
-            timezone: America/Los_Angeles
-
-        If no timezone is given, or it's invalid, time will be reported in UTC
-        """
-
-        regex = r'^(?:score|karma)\s+--when-(positive|negative)'
-
-        def command(self, bot, comm, groups):
-            kt = bot.factory.loader.db.session.query(KarmaStatsTable)
-            counter = Counter()
-
-            if groups[0] == "positive":
-                karma = kt.filter(KarmaStatsTable.kcount > 0)
-            elif groups[0] == "negative":
-                karma = kt.filter(KarmaStatsTable.kcount < 0)
-
-            for row in karma:
-                hour = row.datetime.hour
-                counter[hour] += row.kcount
-
-            common_hour = (counter.most_common(1)[0][0]
-                           if counter.most_common(1) else None)
-
-            # Title case for when
-            title_case = groups[0][0].upper() + groups[0][1:]
-
-            if common_hour:
-                # Create a datetime object
-                current_time = datetime.now(pytz.utc)
-                # Give it the common_hour
-                current_time = current_time.replace(hour=int(common_hour))
-                # Get the localized common hour
-                hour = self.plugin.tzinfo.normalize(
-                    current_time.astimezone(self.plugin.tzinfo)).hour
-                # Report to the channel
                 bot.reply(
-                    comm,
-                    '%s karma is usually given during the %d:00 hour (%s)' %
-                    (title_case, hour, self.plugin.timezone)
+                    comm, '%s has %d points' % (uen(user.user), user.kcount),
+                    encode=False
                 )
             else:
-                # Inform that no karma of that type has been awarded yet
                 bot.reply(
-                    comm,
-                    '%s karma has been given yet' % title_case
+                    comm, 'No karma for %s ' % uen(thing), encode=False
                 )
 
 
 class KarmaTable(SQLAlchemyBase):
     """
-    Bringing back the classic table so data doesn't need to be dumped
+    Keep track of users karma in a persistant manner
     """
 
     __tablename__ = 'karma'
 
-    # Karma Classic Table
+    # Calling the primary key user, though, really, this can be any string
     user = Column(String, primary_key=True)
     kcount = Column(Integer)
 
-    def __init__(self, user, kcount):
+    def __init__(self, user, kcount=0):
         self.user = user
         self.kcount = kcount
 
 
-class KarmaStatsTable(SQLAlchemyBase):
-    """
-    Keep track of users karma in a persistant manner
-    """
-
-    __tablename__ = 'karmastats'
-
-    # Calling the primary key user, though, really, this can be any string
-    id = Column(Integer, primary_key=True)
-    giver = Column(String)
-    receiver = Column(String)
-    kcount = Column(Integer)
-    datetime = Column(DateTime, default=datetime.utcnow())
-
-    def __init__(self, giver, receiver, kcount):
-        self.giver = giver
-        self.receiver = receiver
-        self.kcount = kcount
+karma = Karma()
